fix livePhoto issue, impl viewGrid delete ability
