+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "immich-mediakit"
-version = "0.1.5"
+version = "0.1.6"
 
 [tool.pyright]
 include = ["src"]
